+++ src/modules/billing/bull-jobs/create-bill-run.js
@@ -1,11 +1,10 @@
+const path = require('path');
 const Bull = require('bull');
-const { get } = require('lodash');
 
 const logger = require('./lib/logger');
 const helpers = require('./lib/helpers');
 
 const { BATCH_ERROR_CODE } = require('../../../lib/models/batch');
-const batchService = require('../services/batch-service');
 
 const JOB_NAME = 'billing.create-bill-run.*';
 
@@ -23,20 +22,6 @@ const publish = (batch, eventId) => queue.add({
   jobId: helpers.createJobId(JOB_NAME, batch)
 });
 
-/**
- * Job handler - creates bill run in charge module
- * @param {Object} job
- * @param {Object} job.batch
- */
-const jobHandler = async job => {
-  logger.logHandling(job);
-  const batchId = get(job, 'data.batch.id');
-  const batch = await batchService.createChargeModuleBillRun(batchId);
-  return {
-    batch
-  };
-};
-
 const completedHandler = async (job, result) => {
   logger.logCompleted(job);
   populateBatchChargeVersionsJob.publish({
@@ -48,11 +33,10 @@ const completedHandler = async (job, result) => {
 const failedHandler = helpers.createFailedHandler(BATCH_ERROR_CODE.failedToCreateBillRun, queue, JOB_NAME);
 
 // Set up queue
-queue.process(jobHandler);
+queue.process(path.join(__dirname, '/processors/create-bill-run.js'));
 queue.on('completed', completedHandler);
 queue.on('failed', failedHandler);
 
-exports.jobHandler = jobHandler;
 exports.failedHandler = failedHandler;
 exports.publish = publish;
 exports.JOB_NAME = JOB_NAME;
