+++ src/modules/billing/bull-jobs/prepare-transactions.js
@@ -1,3 +1,4 @@
+const path = require('path');
 const Bull = require('bull');
 
 const logger = require('./lib/logger');
@@ -5,10 +6,6 @@ const helpers = require('./lib/helpers');
 
 const { BATCH_ERROR_CODE } = require('../../../lib/models/batch');
 const batchService = require('../services/batch-service');
-const billingVolumesService = require('../services/billing-volumes-service');
-const transactionService = require('../services/transactions-service');
-const supplementaryBillingService = require('../services/supplementary-billing-service');
-const repos = require('../../../lib/connectors/repos');
 
 const createChargeJob = require('./create-charge');
 
@@ -24,34 +21,6 @@ const publish = data => queue.add(data, {
   jobId: helpers.createJobId(JOB_NAME, data.batch)
 });
 
-/**
- * Job handler - creates bill run in charge module
- * @param {Object} job
- * @param {Object} job.batch
- */
-const jobHandler = async job => {
-  logger.logHandling(job);
-
-  const batch = await batchService.getBatchById(job.data.batch.id);
-  const billingVolumesForBatch = await billingVolumesService.getVolumesForBatch(batch);
-
-  if (billingVolumesForBatch.length > 0) {
-    await transactionService.updateTransactionVolumes(batch);
-  }
-
-  if (batch.isSupplementary()) {
-    logger.logInfo(job, 'Processing supplementary transactions');
-    await supplementaryBillingService.processBatch(batch.id);
-  }
-
-  const transactions = await repos.billingTransactions.findByBatchId(batch.id);
-
-  return {
-    batch,
-    transactions
-  };
-};
-
 const completedHandler = async (job, result) => {
   logger.logCompleted(job);
 
@@ -59,7 +28,8 @@ const completedHandler = async (job, result) => {
 
   if (transactions.length === 0) {
     logger.logInfo(job, 'empty batch');
-    return batchService.setStatusToEmptyWhenNoTransactions(batch);
+    await batchService.setStatusToEmptyWhenNoTransactions(batch);
+    return;
   }
 
   logger.logInfo(job, `${transactions.length} transactions produced, creating charges...`);
@@ -76,11 +46,10 @@ const completedHandler = async (job, result) => {
 const failedHandler = helpers.createFailedHandler(BATCH_ERROR_CODE.failedToPrepareTransactions, queue, JOB_NAME);
 
 // Set up queue
-queue.process(jobHandler);
+queue.process(path.join(__dirname, '/processors/prepare-transactions.js'));
 queue.on('completed', completedHandler);
 queue.on('failed', failedHandler);
 
-exports.jobHandler = jobHandler;
 exports.failedHandler = failedHandler;
 exports.publish = publish;
 exports.JOB_NAME = JOB_NAME;
