+++ src/modules/billing/bull-jobs/processors/create-charge.js
@@ -1,0 +1,31 @@
+const { get } = require('lodash');
+
+const transactionsService = require('../../services/transactions-service');
+const logger = require('../lib/logger');
+const mappers = require('../../mappers');
+const chargeModuleBillRunConnector = require('../../../../lib/connectors/charge-module/bill-runs');
+
+/**
+ * Job handler - creates bill run in charge module
+ * @param {Object} job
+ * @param {Object} job.batch
+ */
+const jobHandler = async job => {
+  logger.logHandling(job);
+
+  const transactionId = get(job, 'data.transaction.billingTransactionId');
+
+  // Create batch model from loaded data
+  const batch = await transactionsService.getById(transactionId);
+
+  // Map data to charge module transaction
+  const [cmTransaction] = mappers.batch.modelToChargeModule(batch);
+
+  // Create transaction in Charge Module
+  const response = await chargeModuleBillRunConnector.addTransaction(batch.externalId, cmTransaction);
+
+  // Update/remove our local transaction in water.billing_transactions
+  await transactionsService.updateWithChargeModuleResponse(transactionId, response);
+};
+
+module.exports = jobHandler;
